refactor tests & export relay test utils
