+++ __tests__/ReactRelayQueryRenderer-test.tsx
@@ -35,9 +35,7 @@ import { createOperationDescriptor, Network, Observable, ROOT_ID } from 'relay-r
 
 import { ROOT_TYPE } from 'relay-runtime/lib/store/RelayStoreUtils';
 
-import { generateAndCompile } from './TestCompiler';
-import { createMockEnvironment } from './RelayModernEnvironmentMock';
-import { createPersistedStorage } from './Utils';
+import { generateAndCompile, createMockEnvironment, createPersistedStore, createPersistedRecordSource } from '../src-test';
 
 function expectToBeRendered(
     renderSpy,
@@ -417,9 +415,9 @@ describe('ReactRelayQueryRenderer', () => {
                 it('fetches the query once, always renders snapshot returned by fetch', async () => {
                     const fetch = jest.fn().mockReturnValueOnce(response);
                     store = new Store(
-                        new RecordSource({ storage: createPersistedStorage() }),
+                        new RecordSource({ storage: createPersistedRecordSource() }),
                         {
-                            storage: createPersistedStorage(),
+                            storage: createPersistedStore(),
                         },
                         { queryCacheExpirationTime: null },
                     );
@@ -691,9 +689,9 @@ describe('ReactRelayQueryRenderer', () => {
             it('skip loading state when request could be resolved synchronously', async () => {
                 const fetch = () => response;
                 store = new Store(
-                    new RecordSource({ storage: createPersistedStorage() }),
+                    new RecordSource({ storage: createPersistedRecordSource() }),
                     {
-                        storage: createPersistedStorage(),
+                        storage: createPersistedStore(),
                     },
                     { queryCacheExpirationTime: null },
                 );
@@ -734,9 +732,9 @@ describe('ReactRelayQueryRenderer', () => {
                 const error = new Error('Mock Network Error');
                 const fetch: any = () => error;
                 store = new Store(
-                    new RecordSource({ storage: createPersistedStorage() }),
+                    new RecordSource({ storage: createPersistedRecordSource() }),
                     {
-                        storage: createPersistedStorage(),
+                        storage: createPersistedStore(),
                     },
                     { queryCacheExpirationTime: null },
                 );
