--- __tests__/Utils.ts
@@ -1,11 +1,0 @@
-export function createPersistedStorage(clientState = {}) {
-    const state = {};
-    Object.keys(clientState).forEach((key) => (state['relay-records.' + key] = JSON.stringify(clientState[key])));
-    return {
-        getAllKeys: () => Promise.resolve(Object.keys(state)),
-        setItem: (key, value) => Promise.resolve(),
-        removeItem: (key) => Promise.resolve(delete state[key]),
-        getItem: (key) => Promise.resolve(state[key]),
-        getState: () => state,
-    } as any;
-}
