+++ tsconfig-test.json
@@ -1,7 +1,7 @@
 {
     "compilerOptions": {
-        "outDir": "lib",
-        "rootDir": "src",
+        "outDir": "test",
+        "rootDir": "src-test",
         "module": "commonjs",
         "target": "es5",
         "moduleResolution": "node",
@@ -22,7 +22,7 @@
         "test",
         "__tests__",
         "scripts",
-        "src-test"
+        "src"
     ],
     "compileOnSave": true
 }
